requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
open-clip-torch==2.23.0
pymongo==4.6.1
dnspython==2.4.2
python-dotenv==1.0.0